content/html/content/test/test_bug41464.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=41464
-->
<head>
  <title>Test for Bug 41464</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=41464">Mozilla Bug 41464</a>
<p id="display"></p>
<div id="content" style="display: none">
<textarea wrap=cheesecake></textarea>  
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 41464 **/
var textarea = document.getElementById("content").getElementsByTagName("textarea")[0];
is(textarea.wrap, "cheesecake");
is(textarea.hasAttribute("wrap"), true);
is(textarea.getAttribute("wrap"), "cheesecake");

textarea.setAttribute("wrap", "off");
is(textarea.wrap, "off");
is(textarea.hasAttribute("wrap"), true);
is(textarea.getAttribute("wrap"), "off");

textarea.wrap = "soft";
is(textarea.wrap, "soft");
is(textarea.hasAttribute("wrap"), true);
is(textarea.getAttribute("wrap"), "soft");

textarea.wrap = undefined;
is(textarea.wrap, "undefined");
is(textarea.hasAttribute("wrap"), true);
is(textarea.getAttribute("wrap"), "undefined");

textarea.wrap = null;
is(textarea.wrap, "null");
is(textarea.hasAttribute("wrap"), true);
is(textarea.getAttribute("wrap"), "null");

textarea.removeAttribute("wrap");
is(textarea.wrap, "");
is(textarea.hasAttribute("wrap"), false);
</script>
</pre>
</body>
</html>
